+++ src/lib/storage/table.hpp
@@ -71,8 +71,8 @@ class Table : private Noncopyable {
   std::vector<std::string> _column_names;
   std::vector<std::string> _column_types;
   std::vector<bool> _column_nullable;
-  int32_t _column_count = 0;
-  int32_t _chunk_count = 0;
+  uint16_t _column_count = 0;
+  uint16_t _chunk_count = 0;
   ChunkOffset _target_chunk_size;
 };
